requirements.txt
pip
python-dotenv
pytube
setuptools
tk   
pip install